requirements.txt
django==4.1.5
python-decouple==3.8
psycopg2-binary

django-filter
djangorestframework
markdown
drf-yasg[validation]
djangorestframework-simplejwt
#django-mail-templated
djoser
